backend_drf/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.2.2
cffi==1.17.1
channels==2.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorgram.py==1.2.0
confluent-kafka==2.6.0
constantly==23.10.4
cryptography==43.0.3
daphne==2.5.0
decouple==0.0.7
distlib==0.3.8
Django==5.1.2
djangorestframework==3.15.2
dnspython==2.7.0
eventlet==0.37.0
Faker==30.8.2
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
gevent==24.10.3
greenlet==3.1.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
lxml==5.2.1
MarkupSafe==2.1.5
multidict==6.0.5
mysqlclient==2.2.4
numpy==1.26.4
packaging==25.0
pandas==2.2.1
pika==1.3.2
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
service-identity==24.2.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
timer==0.2.2
twilio==9.0.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.6
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.42.0
yarl==1.9.4
zope.event==5.0
zope.interface==7.1.1
